Alerts right now are working properly. Cleaned redundant Alerts code. Signup/Login functionality works properly

+++ app/src/main/java/com/example/sep4android/viewModel/SetAlertViewModel.java
@@ -49,8 +49,8 @@ public class SetAlertViewModel extends AndroidViewModel {
     ------------------------------------------------------------------
  */
     //  Temperature Alerts/Get/Update/Insert
-    // GETTING THE ALERT INFO!!!!!!!!!!!!!!!!!!!!!
-    public LiveData<TemperatureAlert> getTemperatureAlert(long userId) {
+
+    public LiveData<TemperatureAlert> getTemperatureAlert(int userId) {
         return alertRepository.getTemperatureAlert(userId);
     }
     //  Temperature Alerts/Get/Update/Insert
